backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.13.0
python-dateutil==2.8.2
reportlab==4.0.7
openpyxl==3.1.2
Pillow==10.1.0
